% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.SAG}
\alias{coef.SAG}
\title{Model's weights
Returns the model's weights for each specified lambda.}
\usage{
\method{coef}{SAG}(object, ...)
}
\arguments{
\item{object}{object of class SAG}

\item{...}{Any other pass-through parameters}
}

