% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.SAG_fit}
\alias{coef.SAG_fit}
\title{Model's weights
Returns the model's weights.}
\usage{
\method{coef}{SAG_fit}(object, ...)
}
\arguments{
\item{object}{object of class SAG}

\item{...}{Any other pass-through parameters}
}

