% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_approx_grad.SAG_fit}
\alias{get_approx_grad.SAG_fit}
\title{get approximate gradient
Returns the models approximate gradient}
\usage{
\method{get_approx_grad}{SAG_fit}(object, ...)
}
\arguments{
\item{object}{of class SAG_fit}

\item{...}{Any other pass-through parameters}
}

