% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_cost.SAG}
\alias{get_cost.SAG}
\title{Model's cost
Returns the model's true cost for each specified lambda.}
\usage{
\method{get_cost}{SAG}(object, X, y, ...)
}
\arguments{
\item{object}{object of class SAG}

\item{X}{Matrix of samples}

\item{y}{Matrix of targets}

\item{...}{Any other pass-through parameters}
}

