% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{get_cost.SAG_fit}
\alias{get_cost.SAG_fit}
\title{Model's costs
Returns the model's true cost.}
\usage{
\method{get_cost}{SAG_fit}(object, X, y, ...)
}
\arguments{
\item{object}{object of class SAG_fit}

\item{X}{Matrix of samples}

\item{y}{Matrix of targets}

\item{...}{Any other pass-through parameters}
}
\description{
Model's costs
Returns the model's true cost.
}

