% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_grad.SAG}
\alias{get_grad.SAG}
\title{Model's gradients
Returns gradients of model for each lambda.}
\usage{
\method{get_grad}{SAG}(object, X, y, ...)
}
\arguments{
\item{object}{object of class SAG}

\item{X}{Matrix of samples}

\item{y}{Matrix of targets}

\item{...}{Any other pass-through parameters}
}

