% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigoptim.R
\docType{data}
\name{GlmnetExamples}
\alias{GlmnetExamples}
\title{GlmnetExamples
Adapted from GLMNET}
\format{WRITEME!}
\source{
WRITEME!
}
\usage{
GlmnetExamples
}
\description{
GlmnetExamples
Adapted from GLMNET
}
\keyword{datasets}

