% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_C_losses.R
\name{make_c_loss}
\alias{make_c_loss}
\title{Make user supplied C loss and gradient}
\usage{
make_c_loss(src, loss_name, grad_name)
}
\arguments{
\item{src}{C source code as character string}

\item{loss_name}{Loss function name}

\item{grad_name}{Gradient function name}
}

