% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigoptim.R
\docType{data}
\name{mini.covtype.libsvm}
\alias{mini.covtype.libsvm}
\title{mini.covtype.libsvm
Random sample of 3000 points from covtype.libsvm}
\format{WRITEME!}
\source{
WRITEME!
mini.covtype.libsvm
}
\usage{
mini.covtype.libsvm
}
\description{
mini.covtype.libsvm
Random sample of 3000 points from covtype.libsvm
}
\keyword{datasets}

