% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigoptim.R
\docType{data}
\name{rcv1_train}
\alias{rcv1_train}
\title{mini.covtype.libsvm
Random sample of 3000 points from covtype.libsvm}
\format{WRITEME!}
\source{
WRITEME!
mini.covtype.libsvm
rcv1_train

WRITEME!
}
\usage{
rcv1_train
}
\description{
mini.covtype.libsvm
Random sample of 3000 points from covtype.libsvm
}
\keyword{datasets}

