% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainer.R
\name{sag_constant}
\alias{sag_constant}
\title{SAG with constant step size for logistic regression}
\usage{
sag_constant(X, y, lambda = 0, maxiter = NULL, wInit = NULL,
  stepSize = NULL, iVals = NULL, d = NULL, g = NULL, covered = NULL)
}
\arguments{
\item{X}{matrix of examples n X m. Will be transposed when .Call}

\item{y}{matrix of targets n x 1}

\item{lambda}{scalar regularisation parameter}

\item{stepSize}{scalar constant step size}

\item{iVals}{Sequenc of examples to choose maxiter X 1}

\item{d}{initial approximation of average gradient p X 1 (should be the sum}

\item{g}{previous derivativers of loss.}

\item{covered}{whether the example has been visited. n X 1}

\item{w}{vector of weights m X 1}

\item{sequence}{of examples to choose}
}
\value{
list of results
}
\description{
let n be the number of example. p the number of features
}
\author{
Mohamed Ishmael Diwan Belghazi
}

