% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainer.R
\name{sag_fit}
\alias{sag_fit}
\title{Stochastic Average Gradient Fit}
\usage{
sag_fit(X, y, lambda = 0, maxiter = NULL, w = NULL, stepSize = NULL,
  stepSizeType = 1, Li = NULL, Lmax = NULL, increasing = TRUE,
  iVals = NULL, d = NULL, g = NULL, covered = NULL,
  standardize = TRUE, randVals = randVals, tol = 0.001,
  model = "binomial", fit_alg = "constant", ...)
}
\value{
object of class SAG_fit
}
\description{
Stochastic Average Gradient Fit
}

