% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainer.R
\name{sag_fit}
\alias{sag_fit}
\title{Stochastic Average Gradient Fit}
\usage{
sag_fit(X, y, lambda = 0, maxiter = NULL, w = NULL, alpha = NULL,
  stepSizeType = 1, Li = NULL, Lmax = NULL, increasing = TRUE,
  d = NULL, g = NULL, covered = NULL, standardize = FALSE,
  tol = 0.001, family = "binomial", fit_alg = "constant",
  monitor = FALSE, ...)
}
\value{
object of class SAG_fit
}
\description{
Stochastic Average Gradient Fit
}

