% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainer.R
\name{sag_fit}
\alias{sag_fit}
\title{Stochastic Average Gradient}
\usage{
sag_fit(X, y, lambda = 0, maxiter = NULL, w = NULL, alpha = NULL,
  stepSizeType = 1, Li = NULL, Lmax = NULL, increasing = TRUE,
  d = NULL, g = NULL, covered = NULL, standardize = FALSE,
  tol = 0.001, family = "binomial", fit_alg = "constant",
  monitor = FALSE, user_loss_function = NULL, ...)
}
\arguments{
\item{X}{Matrix, possibly sparse of features.}

\item{y}{Matrix of targets.}

\item{lambda}{Scalar. L2 regularization parameter.}

\item{maxiter}{Maximum number of iterations.}

\item{w}{Matrix of weights.}

\item{alpha}{constant step-size. Used only when fit_alg == "constant"}

\item{stepSizeType}{scalar default is 1 to use 1/L, set to 2 to use 2/(L + n*myu). Only used when fit_alg="linesearch"}

\item{Li}{Scalar or Matrix.Initial individual Lipschitz approximation.}

\item{Lmax}{Initial global Lipschitz approximation.}

\item{increasing}{Boolean. TRUE allows for both increase and decrease of lipschitz coefficients. False allows only decrease.}

\item{d}{Initial approximation of cost function gradient.}

\item{g}{Initial approximation of individual losses gradient.}

\item{covered}{Matrix of covered samples.}

\item{standardize}{Boolean. Scales the data if True}

\item{tol}{Real. Miminal required approximate gradient norm before convergence.}

\item{family}{One of "binomial", "gaussian", "exponential" or "poisson"}

\item{fit_alg}{One of "constant", "linesearch" (default), or "adaptive"}

\item{monitor}{Boolean. If TRUE returns matrix of weights after each effective pass through the dataset.}

\item{...}{Any other pass-through parameters.}
}
\value{
object of class SAG_fit
}
\description{
Stochastic Average Gradient
}

